LeAudio: Improve handling remote device audio locations
With this change we correctly distinguish devices with Mono audio location (0),
from devices which do not support any audio location in a paritcular direction.
This is important for devices which are one direction only - speakers
only or microphone only.

Bug: 388499616
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Flag: Exempt: day to day fix

+++ system/gd/rust/topshim/src/profiles/le_audio.rs
@@ -194,8 +194,8 @@ pub mod ffi {
         fn le_audio_audio_conf_callback(
             direction: u8,
             group_id: i32,
-            snk_audio_location: u32,
-            src_audio_location: u32,
+            snk_audio_location: i64,
+            src_audio_location: i64,
             avail_cont: u16,
         );
         fn le_audio_sink_audio_location_available_callback(
@@ -520,7 +520,7 @@ cb_variant!(LeAudioClientCb,
 
 cb_variant!(LeAudioClientCb,
             le_audio_audio_conf_callback -> LeAudioClientCallbacks::AudioConf,
-            u8, i32, u32, u32, u16);
+            u8, i32, i64, i64, u16);
 
 cb_variant!(LeAudioClientCb,
             le_audio_sink_audio_location_available_callback -> LeAudioClientCallbacks::SinkAudioLocationAvailable,
